requirements.txt
numpy==1.26.4
tensorflow==2.15.1
pyomo==6.7.2
scikit-learn==1.5.0
seaborn==0.13.2
pandas==2.2.2
omlt==1.1
liac-arff==2.4.0
